feat:module client and vehicles finished

+++ src/main/java/com/main/entities/Servicios.java
@@ -22,5 +22,5 @@ public class Servicios {
     private String nombre;
     private Double costo;
     private String descripcion;
-    private Double tiempoEstimado;
+    private Integer tiempoEstimado;
 }
